refactor(create-deal): reduce component to pure UI orchestration
Step 4/4 of #70. Extracts init effect into hooks/use-create-deal-init.ts,
form state and derived values into hooks/use-create-deal-form.ts, and
JSX body into app/create-deal/components/create-deal-form-body.tsx.
create-deal-content.tsx reduced from 546 to 108 lines.

Adds input validation to submit() for userId, signerAddress,
MERCATO_PLATFORM_ADDRESS, and USDC_TRUSTLINE. Fixes 3 TS type
mismatches from the extraction (userId nullability, selectedProduct
undefined case, updateFormData keyof signature).

Verified: tsc --noEmit error count back to 10 (baseline).
